Rename containers for consistency
